Function References

Methods

ImageDraw.draw — Function

img_new = draw(img, drawable, color)
img_new = draw(img, [drawable], [color])

Draws the drawable object on a copy of image img using color color. Can also draw multiple Drawable objects when passed as a AbstractVector{Drawable} with corresponding colors in [color]

ImageDraw.draw! — Function

img = draw!(img, drawable, color)
img = draw!(img, drawable)

Draws drawable on img using color color which defaults to oneunit(eltype(img))

img = draw!(img, [drawable], [color])
img = draw!(img, [drawable] ,color)
img = draw!(img, [drawable])

Draws all objects in [drawable] in the given order on img using corresponding colors from [color] which defaults to oneunit(eltype(img)) If only a single color color is specified then all objects will be colored with that color.

ImageDraw.bresenham — Function

res = bresenham(img, y0, x0, y1, x1, color)

Method to generate a line profile from (x0,y0) to (x1,y1) of a 2d image using Bresenham's algorithm.

ImageDraw.xiaolin_wu — Function

res = xiaolin_wu(img, x0, y0, x1, x2, color)

Method to generate a line profile from (x0,y0) to (x1,y1) of a 2d image using Xiaolin Wu line algorithm.